+++ src/factor_instance_provider/factor_instance_provider.rs
@@ -721,75 +721,6 @@ impl FactorInstanceProvider {
 }
 
 impl FactorInstanceProvider {
-    pub async fn securify_with_address<E: IsEntity + std::hash::Hash + std::cmp::Eq>(
-        &self,
-        address: &E::Address,
-        matrix: MatrixOfFactorSources,
-        profile: &mut Profile,
-    ) -> Result<SecurifiedEntityControl> {
-        let entity = profile.entity_by_address::<E>(address)?;
-        self.securify(&entity, &matrix, profile).await
-    }
-
-    pub async fn securify<E: IsEntity>(
-        &self,
-        entity: &E,
-        matrix: &MatrixOfFactorSources,
-        profile: &mut Profile,
-    ) -> Result<SecurifiedEntityControl> {
-        let entity_kind = E::kind();
-        let network_id = entity.address().network_id();
-        let key_kind = CAP26KeyKind::TransactionSigning;
-
-        let requests = matrix
-            .clone()
-            .all_factors()
-            .into_iter()
-            .map(|factor_source| {
-                DerivationRequest::securify(
-                    entity_kind,
-                    key_kind,
-                    factor_source.factor_source_id(),
-                    network_id,
-                )
-            })
-            .collect::<IndexSet<_>>();
-
-        let derived_factors_map = self.provide_factor_instances(profile, requests).await?;
-
-        let derived_factors = derived_factors_map
-            .values()
-            .cloned()
-            .collect::<IndexSet<_>>();
-
-        let matrix = MatrixOfFactorInstances::fulfilling_matrix_of_factor_sources_with_instances(
-            derived_factors,
-            matrix.clone(),
-        )?;
-
-        let component_metadata = ComponentMetadata::new(matrix.clone());
-
-        let securified_entity_control = SecurifiedEntityControl::new(
-            matrix,
-            AccessController {
-                address: AccessControllerAddress::new(entity.entity_address()),
-                metadata: component_metadata,
-            },
-        );
-
-        profile.update_entity(E::new(
-            entity.name(),
-            entity.entity_address(),
-            EntitySecurityState::Securified(securified_entity_control.clone()),
-        ));
-
-        let gateway = self.gateway.clone();
-        gateway
-            .set_securified_entity(securified_entity_control.clone(), entity.address())
-            .await?;
-        Ok(securified_entity_control)
-    }
-
     pub async fn provide_genesis_factor_for<'p>(
         &self,
         factor_source_id: FactorSourceIDFromHash,
@@ -843,5 +774,88 @@ impl FactorInstanceProvider {
                 self.fulfill(profile, fulfillable, unfulfillable).await
             }
         }
+    }
+}
+
+#[cfg(test)]
+mod securify_tests {
+
+    use super::*;
+
+    #[actix_rt::test]
+    async fn derivation_path_is_never_same_after_securified() {
+        let all_factors = HDFactorSource::all();
+        let a = &Account::unsecurified_mainnet(
+            "A0",
+            HierarchicalDeterministicFactorInstance::mainnet_tx(
+                CAP26EntityKind::Account,
+                HDPathComponent::unsecurified_hardening_base_index(0),
+                FactorSourceIDFromHash::fs0(),
+            ),
+        );
+        let b = &Account::unsecurified_mainnet(
+            "A1",
+            HierarchicalDeterministicFactorInstance::mainnet_tx(
+                CAP26EntityKind::Account,
+                HDPathComponent::unsecurified_hardening_base_index(1),
+                FactorSourceIDFromHash::fs0(),
+            ),
+        );
+
+        let mut profile = Profile::new(all_factors.clone(), [a, b], []);
+        let matrix = MatrixOfFactorSources::new([fs_at(0)], 1, []);
+
+        let interactors = Arc::new(TestDerivationInteractors::default());
+        let gateway = Arc::new(TestGateway::default());
+
+        let factor_instance_provider = FactorInstanceProvider::new(
+            gateway.clone(),
+            interactors,
+            Arc::new(InMemoryPreDerivedKeysCache::default()),
+        );
+
+        let b_sec = factor_instance_provider
+            .securify(b, &matrix, &mut profile)
+            .await
+            .unwrap();
+
+        assert_eq!(
+            b_sec
+                .matrix
+                .all_factors()
+                .clone()
+                .into_iter()
+                .map(|f| f.derivation_path().index)
+                .collect::<HashSet<_>>()
+                .len(),
+            1
+        );
+
+        assert_eq!(
+            b_sec
+                .matrix
+                .all_factors()
+                .clone()
+                .into_iter()
+                .map(|f| f.derivation_path().index)
+                .collect::<HashSet<_>>(),
+            HashSet::just(HDPathComponent::securifying_base_index(0))
+        );
+
+        let a_sec = factor_instance_provider
+            .securify(a, &matrix, &mut profile)
+            .await
+            .unwrap();
+
+        assert_eq!(
+            a_sec
+                .matrix
+                .all_factors()
+                .clone()
+                .into_iter()
+                .map(|f| f.derivation_path().index)
+                .collect::<HashSet<_>>(),
+            HashSet::just(HDPathComponent::securifying_base_index(1))
+        );
     }
 }
