+++ src/factor_instance_provider/mod.rs
@@ -1,4 +1,13 @@
+mod derivation_request;
 #[allow(clippy::module_inception)]
 mod factor_instance_provider;
 
+mod gateway;
+mod profile_extensions;
+mod securify;
+
+pub use derivation_request::*;
 pub use factor_instance_provider::*;
+pub use gateway::*;
+pub use profile_extensions::*;
+pub use securify::*;
