+++ src/factor_instance_provider/securify/mod.rs
@@ -1,0 +1,4 @@
+#[allow(clippy::module_inception)]
+mod securify;
+
+pub use securify::*;
