+++ src/recovery/recovery.rs
@@ -2,277 +2,6 @@ use std::{hash::Hash, ops::Add, sync::RwLock};
 
 use crate::prelude::*;
 
-impl Profile {
-    async fn new_entity<E: IsEntity + std::fmt::Debug + std::hash::Hash + Eq>(
-        &mut self,
-        network_id: NetworkID,
-        name: impl AsRef<str>,
-        factor_source_id: FactorSourceIDFromHash,
-        factor_instance_provider: &FactorInstanceProvider,
-    ) -> Result<E> {
-        assert!(self
-            .factor_sources
-            .iter()
-            .map(|f| f.factor_source_id())
-            .contains(&factor_source_id));
-
-        let genesis_factor = factor_instance_provider
-            .provide_genesis_factor_for(factor_source_id, E::kind(), network_id, self)
-            .await?;
-
-        let address = E::Address::by_hashing(network_id, genesis_factor.clone());
-
-        let entity = E::new(
-            name,
-            address,
-            EntitySecurityState::Unsecured(genesis_factor),
-        );
-
-        let erased = Into::<AccountOrPersona>::into(entity.clone());
-
-        match erased {
-            AccountOrPersona::AccountEntity(account) => {
-                self.accounts.insert(account.entity_address(), account);
-            }
-            AccountOrPersona::PersonaEntity(persona) => {
-                self.personas.insert(persona.entity_address(), persona);
-            }
-        };
-
-        Ok(entity)
-    }
-
-    pub async fn new_account(
-        &mut self,
-        network_id: NetworkID,
-        name: impl AsRef<str>,
-        factor_source_id: FactorSourceIDFromHash,
-        factor_instance_provider: &FactorInstanceProvider,
-    ) -> Result<Account> {
-        self.new_entity(network_id, name, factor_source_id, factor_instance_provider)
-            .await
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct OnChainEntityUnsecurified {
-    address: AddressOfAccountOrPersona,
-    owner_keys: Vec<PublicKeyHash>,
-}
-impl OnChainEntityUnsecurified {
-    pub fn new(
-        address: impl Into<AddressOfAccountOrPersona>,
-        owner_keys: Vec<PublicKeyHash>,
-    ) -> Self {
-        Self {
-            address: address.into(),
-            owner_keys,
-        }
-    }
-
-    pub fn owner_keys(&self) -> HashSet<PublicKeyHash> {
-        self.owner_keys.iter().cloned().collect()
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct OnChainEntitySecurified {
-    address: AddressOfAccountOrPersona,
-    access_controller: AccessController,
-    owner_keys: Vec<PublicKeyHash>,
-}
-
-impl OnChainEntitySecurified {
-    pub fn new(
-        address: impl Into<AddressOfAccountOrPersona>,
-        access_controller: AccessController,
-        owner_keys: Vec<PublicKeyHash>,
-    ) -> Self {
-        Self {
-            address: address.into(),
-            access_controller,
-            owner_keys,
-        }
-    }
-    pub fn owner_keys(&self) -> HashSet<PublicKeyHash> {
-        self.owner_keys.iter().cloned().collect()
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash, EnumAsInner)]
-pub enum OnChainEntityState {
-    Unsecurified(OnChainEntityUnsecurified),
-    Securified(OnChainEntitySecurified),
-}
-impl OnChainEntityState {
-    fn unsecurified(unsecurified: OnChainEntityUnsecurified) -> Self {
-        Self::Unsecurified(unsecurified)
-    }
-
-    pub fn unsecurified_with(
-        address: impl Into<AddressOfAccountOrPersona>,
-        owner_key: PublicKeyHash,
-    ) -> Self {
-        Self::unsecurified(OnChainEntityUnsecurified::new(address, vec![owner_key]))
-    }
-
-    fn securified(securified: OnChainEntitySecurified) -> Self {
-        Self::Securified(securified)
-    }
-
-    pub fn securified_with(
-        address: impl Into<AddressOfAccountOrPersona>,
-        access_controller: AccessController,
-        owner_keys: Vec<PublicKeyHash>,
-    ) -> Self {
-        Self::securified(OnChainEntitySecurified::new(
-            address,
-            access_controller.clone(),
-            owner_keys,
-        ))
-    }
-}
-
-impl OnChainEntityState {
-    #[allow(unused)]
-    fn address(&self) -> AddressOfAccountOrPersona {
-        match self {
-            OnChainEntityState::Unsecurified(account) => account.address.clone(),
-            OnChainEntityState::Securified(account) => account.address.clone(),
-        }
-    }
-
-    fn owner_keys(&self) -> HashSet<PublicKeyHash> {
-        match self {
-            OnChainEntityState::Unsecurified(account) => account.owner_keys(),
-            OnChainEntityState::Securified(account) => account.owner_keys(),
-        }
-    }
-}
-
-#[async_trait::async_trait]
-pub trait GatewayReadonly: Sync + Send {
-    async fn is_key_hash_known(&self, hash: PublicKeyHash) -> Result<bool>;
-    async fn query_public_key_hash_is_known(
-        &self,
-        hashes: IndexSet<PublicKeyHash>,
-    ) -> Result<HashMap<PublicKeyHash, bool>> {
-        let mut is_known_map = HashMap::<PublicKeyHash, bool>::new();
-        for hash in hashes.into_iter() {
-            let is_known = self.is_key_hash_known(hash.clone()).await?;
-            is_known_map.insert(hash, is_known);
-        }
-        Ok(is_known_map)
-    }
-
-    async fn get_entity_addresses_of_by_public_key_hashes(
-        &self,
-        hashes: HashSet<PublicKeyHash>,
-    ) -> Result<HashMap<PublicKeyHash, HashSet<AddressOfAccountOrPersona>>>;
-
-    async fn get_on_chain_entity(
-        &self,
-        address: AddressOfAccountOrPersona,
-    ) -> Result<Option<OnChainEntityState>>;
-
-    async fn get_on_chain_account(
-        &self,
-        account_address: &AccountAddress,
-    ) -> Result<Option<OnChainEntityState>> {
-        self.get_on_chain_entity(account_address.clone().into())
-            .await
-    }
-
-    async fn get_owner_key_hashes(
-        &self,
-        address: AddressOfAccountOrPersona,
-    ) -> Result<Option<HashSet<PublicKeyHash>>> {
-        let on_chain_account = self.get_on_chain_entity(address).await?;
-        return Ok(on_chain_account.map(|account| account.owner_keys().clone()));
-    }
-
-    async fn is_securified(&self, address: AddressOfAccountOrPersona) -> Result<bool> {
-        let entity = self.get_on_chain_entity(address).await?;
-        Ok(entity.map(|x| x.is_securified()).unwrap_or(false))
-    }
-}
-
-#[async_trait::async_trait]
-pub trait Gateway: GatewayReadonly {
-    async fn simulate_network_activity_for(&self, owner: AddressOfAccountOrPersona) -> Result<()>;
-
-    async fn set_securified_entity(
-        &self,
-        securified: SecurifiedEntityControl,
-        owner: AddressOfAccountOrPersona,
-    ) -> Result<()>;
-
-    async fn set_securified_account(
-        &self,
-        securified: SecurifiedEntityControl,
-        owner: &AccountAddress,
-    ) -> Result<()> {
-        self.set_securified_entity(securified, owner.clone().into())
-            .await
-    }
-
-    async fn set_securified_persona(
-        &self,
-        securified: SecurifiedEntityControl,
-        owner: &IdentityAddress,
-    ) -> Result<()> {
-        self.set_securified_entity(securified, owner.clone().into())
-            .await
-    }
-}
-
-pub const DERIVATION_INDEX_BATCH_SIZE: HDPathValue = 50;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct UncoveredEntity {
-    pub on_chain: OnChainEntityState,
-    pub key_hash_to_factor_instances:
-        HashMap<PublicKeyHash, HierarchicalDeterministicFactorInstance>,
-}
-impl UncoveredEntity {
-    pub fn new(
-        on_chain: OnChainEntityState,
-        key_hash_to_factor_instances: HashMap<
-            PublicKeyHash,
-            HierarchicalDeterministicFactorInstance,
-        >,
-    ) -> Self {
-        Self {
-            on_chain,
-            key_hash_to_factor_instances,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct EntityRecoveryOutcome<E>
-where
-    E: IsEntity + Hash + Eq,
-{
-    pub recovered_unsecurified: IndexSet<E>,
-    pub recovered_securified: IndexSet<E>,
-    pub unrecovered: Vec<UncoveredEntity>, // want `IndexSet` but is not `Hash`
-}
-
-impl<E: IsEntity + Hash + Eq> EntityRecoveryOutcome<E> {
-    pub fn new(
-        recovered_unsecurified: impl IntoIterator<Item = E>,
-        recovered_securified: impl IntoIterator<Item = E>,
-        unrecovered: impl IntoIterator<Item = UncoveredEntity>,
-    ) -> Self {
-        Self {
-            recovered_unsecurified: recovered_unsecurified.into_iter().collect(),
-            recovered_securified: recovered_securified.into_iter().collect(),
-            unrecovered: unrecovered.into_iter().collect(),
-        }
-    }
-}
-
 /// A first implementation of Recovery of Securified entities, working POC using
 /// Entity Indexing heuristics - `CEI - strategy 1` - Canonical Entity Indexing,
 /// as described in [doc].
@@ -560,146 +289,6 @@ pub async fn recover_personas(
         gateway,
     )
     .await
-}
-
-#[cfg(test)]
-pub struct TestGateway {
-    /// contains only current state for each entity
-    entities: RwLock<HashMap<AddressOfAccountOrPersona, OnChainEntityState>>,
-
-    /// contains historic state, we only ever add to this set, never remove.
-    known_hashes: RwLock<HashSet<PublicKeyHash>>,
-}
-#[cfg(test)]
-impl Default for TestGateway {
-    fn default() -> Self {
-        Self {
-            known_hashes: RwLock::new(HashSet::new()),
-            entities: RwLock::new(HashMap::new()),
-        }
-    }
-}
-
-#[cfg(test)]
-impl TestGateway {
-    #[allow(unused)]
-    pub fn debug_print(&self) {
-        println!(
-            "⛩️ known_hashes: {:?}",
-            self.known_hashes.try_read().unwrap()
-        );
-        println!("⛩️ entities: {:?}", self.entities.try_read().unwrap().keys());
-    }
-}
-
-#[cfg(test)]
-#[async_trait::async_trait]
-impl GatewayReadonly for TestGateway {
-    async fn is_key_hash_known(&self, hash: PublicKeyHash) -> Result<bool> {
-        let is_known = self.known_hashes.try_read().unwrap().contains(&hash);
-        Ok(is_known)
-    }
-    async fn get_entity_addresses_of_by_public_key_hashes(
-        &self,
-        hashes: HashSet<PublicKeyHash>,
-    ) -> Result<HashMap<PublicKeyHash, HashSet<AddressOfAccountOrPersona>>> {
-        let entities = self.entities.try_read().unwrap();
-        let states = entities.values();
-
-        Ok(hashes
-            .iter()
-            .filter_map(|k| {
-                // N.B. we want this to always be single element (Axiom 1).
-                let mut entities_references_hash = HashSet::<AddressOfAccountOrPersona>::new();
-                for state in states.clone().filter(|x| x.owner_keys().contains(k)) {
-                    entities_references_hash.insert(state.address());
-                }
-                if entities_references_hash.is_empty() {
-                    None
-                } else {
-                    Some((k.clone(), entities_references_hash))
-                }
-            })
-            .collect::<HashMap<PublicKeyHash, HashSet<AddressOfAccountOrPersona>>>())
-    }
-
-    async fn get_on_chain_entity(
-        &self,
-        address: AddressOfAccountOrPersona,
-    ) -> Result<Option<OnChainEntityState>> {
-        Ok(self.entities.try_read().unwrap().get(&address).cloned())
-    }
-}
-#[cfg(test)]
-impl TestGateway {
-    async fn assert_not_securified(&self, address: &AddressOfAccountOrPersona) -> Result<()> {
-        let is_already_securified = self.is_securified(address.clone()).await?;
-        assert!(
-            !is_already_securified,
-            "Cannot unsecurify an already securified entity"
-        );
-        Ok(())
-    }
-
-    fn contains(&self, address: &AddressOfAccountOrPersona) -> bool {
-        self.entities.try_read().unwrap().contains_key(address)
-    }
-}
-
-#[cfg(test)]
-#[async_trait::async_trait]
-impl Gateway for TestGateway {
-    async fn simulate_network_activity_for(&self, owner: AddressOfAccountOrPersona) -> Result<()> {
-        self.assert_not_securified(&owner).await?;
-
-        let owner_key = owner.public_key_hash();
-
-        if self.contains(&owner) || self.known_hashes.try_read().unwrap().contains(&owner_key) {
-            panic!("update not supported")
-        } else {
-            self.entities.try_write().unwrap().insert(
-                owner.clone(),
-                OnChainEntityState::unsecurified_with(owner, owner_key.clone()),
-            );
-            self.known_hashes.try_write().unwrap().insert(owner_key);
-        }
-        Ok(())
-    }
-
-    async fn set_securified_entity(
-        &self,
-        securified: SecurifiedEntityControl,
-        owner: AddressOfAccountOrPersona,
-    ) -> Result<()> {
-        self.assert_not_securified(&owner).await?;
-
-        let owner_keys = securified
-            .matrix
-            .all_factors()
-            .iter()
-            .map(|f| f.public_key_hash())
-            .collect_vec();
-
-        if self.contains(&owner) {
-            self.entities.try_write().unwrap().remove(&owner);
-        }
-
-        self.known_hashes
-            .try_write()
-            .unwrap()
-            .extend(owner_keys.clone());
-
-        self.entities.try_write().unwrap().insert(
-            owner.clone(),
-            OnChainEntityState::securified_with(
-                owner,
-                securified.access_controller.clone(),
-                owner_keys,
-            ),
-        );
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
