--- src/securify/securify_entity.rs
@@ -1,84 +1,0 @@
-use crate::{factor_instance_provider, prelude::*};
-
-#[cfg(test)]
-mod securify_tests {
-
-    use super::*;
-
-    #[actix_rt::test]
-    async fn derivation_path_is_never_same_after_securified() {
-        let all_factors = HDFactorSource::all();
-        let a = &Account::unsecurified_mainnet(
-            "A0",
-            HierarchicalDeterministicFactorInstance::mainnet_tx(
-                CAP26EntityKind::Account,
-                HDPathComponent::unsecurified_hardening_base_index(0),
-                FactorSourceIDFromHash::fs0(),
-            ),
-        );
-        let b = &Account::unsecurified_mainnet(
-            "A1",
-            HierarchicalDeterministicFactorInstance::mainnet_tx(
-                CAP26EntityKind::Account,
-                HDPathComponent::unsecurified_hardening_base_index(1),
-                FactorSourceIDFromHash::fs0(),
-            ),
-        );
-
-        let mut profile = Profile::new(all_factors.clone(), [a, b], []);
-        let matrix = MatrixOfFactorSources::new([fs_at(0)], 1, []);
-
-        let interactors = Arc::new(TestDerivationInteractors::default());
-        let gateway = Arc::new(TestGateway::default());
-
-        let factor_instance_provider = FactorInstanceProvider::new(
-            gateway.clone(),
-            interactors,
-            Arc::new(InMemoryPreDerivedKeysCache::default()),
-        );
-
-        let b_sec = factor_instance_provider
-            .securify(b, &matrix, &mut profile)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            b_sec
-                .matrix
-                .all_factors()
-                .clone()
-                .into_iter()
-                .map(|f| f.derivation_path().index)
-                .collect::<HashSet<_>>()
-                .len(),
-            1
-        );
-
-        assert_eq!(
-            b_sec
-                .matrix
-                .all_factors()
-                .clone()
-                .into_iter()
-                .map(|f| f.derivation_path().index)
-                .collect::<HashSet<_>>(),
-            HashSet::just(HDPathComponent::securifying_base_index(0))
-        );
-
-        let a_sec = factor_instance_provider
-            .securify(a, &matrix, &mut profile)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            a_sec
-                .matrix
-                .all_factors()
-                .clone()
-                .into_iter()
-                .map(|f| f.derivation_path().index)
-                .collect::<HashSet<_>>(),
-            HashSet::just(HDPathComponent::securifying_base_index(1))
-        );
-    }
-}
